INCOMPLETE this is incomplete work from Feb 2016
This is part of a partially working plugin for GDB to allow
breakpoints and stack unwinding in the generated Psyco JIT code,
in theory you should be able to break on Python functions and code
locations etc, but it doesn't work properly in most cases.
it requires the patched fork of GDB on my GitHub.
I think I was trying to understand how to deal with inlined functions.