Cleanup for Py3
Removes a lot of the unneeded/unused funcationality from Psyco. This
should make it much easier to move to Py3.

Most of the code removed was related to debugging which can now be done
with nianalyser or dynamic profiling for code compilation which is not
an immediate goal for Ni.

Compact object is broken! But it needed rethinking anyway to make it
useful.