Start of Python3
Should work on x64 Linux + Python 3.7

Lots of hackery but its now at a point where it will compile and
run a very simple functions:

import ni
def func(a, b):
    return a + b
func(1, 3) # or other types

The mergepoints has all been stripped out and replaced with a very
simple version that will handle the above function. The mergepoints
code was very complex and I don't think we need or want to keep the
early deletion of vars, but I'm not 100% sure it is what I think.

Many optimization have been removed, some would be good to put back
with Python3 versions, but requires analysis.

Compact object has been removed mainly because I don't understand
it so wasn't sure how to port it. It is very fast when used so
analysis should be done around if or how to reinstroduce it. Perhaps
objects could be optimized in some other way?

Much of the remaining code is untested and will required further
work to support Python 3 or to make sure it is doing the right
thing for Python 3. Python 2 behaviour still exists in the
compiler and needs to be changed for Py3.